Task 1 completed.
Using scanner to get user inputs.
The program asks for the user array size.
Then asks for the array elements based on the array size.
A for loop runs through the elements on the array to check if the value is duplicate. The outer loop hold the number to be compared, the inner loop is the comparison.
A boolean hold the value isDuplicate, if the value is false, it measn that the loop did not found any duplicates and send the correct message to the user.