Multiplication table
Multiplication table updated:

First we use scanner to get the side of the multiplication table from the user.
Then we declare a 2d array with [][] and initialize the size of the array with the input from the user.
We will use two variables to hold a number and the multiplier.
The outer loop will deal with the rows of the matrix only, increasing the num variable at the end and reseting the multiplier to 1;
The inner loop will deal with the multiplication and add the result to the right index on the array using the values of i and j to represent the index.

Added an if condition for formatting and readability of the matrix table.