Input validation finalized
On the CalculatorCA2 class::
Used try and catch for the methods to catch when the user is not entering a integer.
For the matrix sizes inputs I used a specific try and catch, so that if the error is caught earlier, it stops the function and gives a specific message.
I also implemented the variable "error" this will hold a boolean value, which will be returned when a error is found.
The error boolean will be used on a conditional at the end of each function. This way we can control what is being sent to the main java file, if a null value is sent to the main file we have a condition handling that error on the correct switch statement.

Additional to that a if condition at the end of the code is being used to control if the user wants to use the calculator or if they want to stop it.